Mejora estetica bottom bar y estetica cards de operator

+++ app/src/main/java/com/endfield/talosIIarchive/domain/repositoty/OperatorRepository.kt
@@ -6,7 +6,7 @@ import com.endfield.talosIIarchive.domain.models.Weapon
 
 interface OperatorRepository {
     suspend fun getAllOperators(): List<Operator>
-    suspend fun getOperatorById(id: Long): Operator?
+    suspend fun getOperatorById(id: Int): Operator?
 
     // --- AÑADE ESTO ---
     suspend fun getAllWeapons(): List<Weapon>
